[#1] 도메인 예외 검증 추가

+++ src/main/java/cart/application/CartItemService.java
@@ -4,12 +4,15 @@ import cart.dao.CartItemDao;
 import cart.dao.ProductDao;
 import cart.domain.CartItem;
 import cart.domain.Member;
+import cart.domain.Product;
 import cart.dto.CartItemQuantityUpdateRequest;
 import cart.dto.CartItemRequest;
 import cart.dto.CartItemResponse;
+import cart.exception.CartItemException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,11 +31,22 @@ public class CartItemService {
     }
 
     public Long add(Member member, CartItemRequest cartItemRequest) {
-        return cartItemDao.save(new CartItem(member, productDao.getProductById(cartItemRequest.getProductId())));
+        Product product = productDao.getProductById(cartItemRequest.getProductId());
+
+        if (Objects.isNull(product)) {
+            throw new CartItemException.NotFound();
+        }
+
+        return cartItemDao.save(new CartItem(member, product));
     }
 
     public void updateQuantity(Member member, Long id, CartItemQuantityUpdateRequest request) {
         CartItem cartItem = cartItemDao.findById(id);
+
+        if (Objects.isNull(cartItem)) {
+            throw new CartItemException("존재하지 않는 장바구니 상품입니다.");
+        }
+
         cartItem.checkOwner(member);
 
         if (request.getQuantity() == 0) {
@@ -46,6 +60,11 @@ public class CartItemService {
 
     public void remove(Member member, Long id) {
         CartItem cartItem = cartItemDao.findById(id);
+
+        if (Objects.isNull(cartItem)) {
+            throw new CartItemException.CartItemNotExists();
+        }
+
         cartItem.checkOwner(member);
 
         cartItemDao.deleteById(id);
